Refactor into app/domain/infra packages with config extraction and module tests
