+++ config/model_params.py
@@ -1,0 +1,5 @@
+DEFAULT_PYTHAG_EXP = 2
+DEFAULT_LOGISTIC_K = 10
+DEFAULT_LOGISTIC_X0 = 0.5
+DEFAULT_IMPUTED_MODE = "full"
+DEFAULT_IMPUTED_WEIGHT = 1.0
