+++ domain/imputation.py
@@ -1,0 +1,25 @@
+import math
+
+
+def infer_default_scores(games_df, stats):
+    df = games_df.copy()
+    if 'is_imputed' not in df.columns:
+        df['is_imputed'] = df['score1'].isna()
+    else:
+        df['is_imputed'] = df['is_imputed'].fillna(df['score1'].isna()).astype(bool)
+    mask = df['score1'].isna()
+    for idx in df[mask].index:
+        t1, t2 = df.at[idx,'team1'], df.at[idx,'team2']
+        st1, st2 = stats[t1], stats[t2]
+        if st1['games'] and st2['games']:
+            avg1 = (st1['gf']/st1['games'] + st2['ga']/st2['games'])/2
+            avg2 = (st2['gf']/st2['games'] + st1['ga']/st1['games'])/2
+        else:
+            avg1 = avg2 = 1
+        loser_avg, winner_avg = sorted([avg1,avg2])
+        loser_score = int(math.floor(loser_avg + 0.5))
+        winner_score = int(math.floor(winner_avg + 0.5)) + 1
+        df.at[idx,'score1'] = winner_score
+        df.at[idx,'score2'] = loser_score
+        df.at[idx,'is_imputed'] = True
+    return df
