+++ domain/stats.py
@@ -1,0 +1,52 @@
+from collections import defaultdict
+
+
+def compute_stats(games):
+    stats = {}
+    h2h = {}
+
+    def ensure_h2h_record(me, opp):
+        key = (me, opp)
+        if key not in h2h:
+            h2h[key] = {'wins': 0, 'games': 0, 'gf': 0, 'ga': 0}
+        return h2h[key]
+
+    teams = set(games['team1']).union(games['team2'])
+    for t in teams:
+        stats[t] = {'wins':0,'losses':0,'ties':0,'gf':0,'ga':0,'games':0,'opponents':[]}
+    for _,r in games.iterrows():
+        t1,t2,s1,s2 = r.team1, r.team2, int(r.score1), int(r.score2)
+        for me,opp,ms,os in [(t1,t2,s1,s2),(t2,t1,s2,s1)]:
+            stats[me]['gf'] += ms
+            stats[me]['ga'] += os
+            stats[me]['games'] += 1
+            stats[me]['opponents'].append(opp)
+            h2h_record = ensure_h2h_record(me, opp)
+            h2h_record['gf'] += ms
+            h2h_record['ga'] += os
+        if s1>s2:
+            stats[t1]['wins']+=1
+            stats[t2]['losses']+=1
+            ensure_h2h_record(t1, t2)['wins']+=1
+        elif s2>s1:
+            stats[t2]['wins']+=1
+            stats[t1]['losses']+=1
+            ensure_h2h_record(t2, t1)['wins']+=1
+        else:
+            stats[t1]['ties']+=1
+            stats[t2]['ties']+=1
+        ensure_h2h_record(t1, t2)['games']+=1
+        ensure_h2h_record(t2, t1)['games']+=1
+    for t,st in stats.items():
+        tot=st['wins']+st['losses']+st['ties']
+        st['win_pct']=st['wins']/tot if tot else 0
+        st['gd']=st['gf']-st['ga']
+    return stats,h2h
+
+
+def compute_sos(stats):
+    sos={}
+    for t,st in stats.items():
+        opps=st['opponents']
+        sos[t]=sum(stats[o]['win_pct'] for o in opps)/len(opps) if opps else 0
+    return sos
