+++ tests/test_ranking_module.py
@@ -1,0 +1,4 @@
+from domain import ranking
+
+def test_ranking_exports():
+    assert hasattr(ranking, 'compute_pythag')
